Update README.md
The README file is excellent as it is right now. But I find that there may be Windows Users like me who are interested in this program and may try to install it. In my opinion, there should at least be a warning for Windows Users to let these users know that Natls will not work in Windows. This is my suggestion which I implemented at the end of README.md.